+++ src/qcommon/cm_polylib.c
@@ -272,7 +272,7 @@ winding_t   *CopyWinding( winding_t *w ) {
 	winding_t   *c;
 
 	c = AllocWinding( w->numpoints );
-	size = (int)( (winding_t *)0 )->p[w->numpoints];
+	size = (size_t)( (winding_t *)0 )->p[w->numpoints];
 	memcpy( c, w, size );
 	return c;
 }
